# static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "Qwallity App API",
    "version": "1.0.0",
    "title": "Qwallity API Documentation",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "tags": [
    {
      "name": "API",
      "description": "Test API"
    }
  ],
  "servers": [
    {
      "url": "/"
    }
  ],
  "components": {
    "schemas": {
      "Users": {
        "type": "object",
        "properties": {
          "first_name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "role_id": {
            "type": "integer",
            "example": 2
          },
          "account": {
            "type": "integer",
            "example": 100
          }
        }
      },
      "Add_Course": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "body": {
            "type": "string"
          },
          "coursetype": {
            "type": "string"
          },
          "author": {
            "type": "string"
          }
        }
      },
      "Amount": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "number"
          }
        }
      },
      "Update_Course": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "body": {
            "type": "string"
          }
        }
      },
      "Courses": {
        "properties": {
          "count": {
            "type": "integer"
          },
          "result": {
            "type": "array",
            "items": {
              "properties": {
                "id": {
                  "type": "integer"
                },
                "title": {
                  "type": "string"
                }
              },
              "required": [
                "id",
                "title"
              ]
            }
          }
        },
        "required": [
          "count",
          "result"
        ]
      },
      "Id": {
        "type": "integer",
        "properties": {
          "name": {
            "type": "integer"
          }
        }
      },
      "update": {
        "type": "object",
        "properties": {
          "first_name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "role_id": {
            "type": "integer",
            "example": 2
          },
          "account": {
            "type": "number",
            "example": 100
          },
          "completed": {
            "type": "boolean"
          }
        }
      }
    },
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic"
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/login/api": {
      "post": {
        "security": [
          {
            "basicAuth": []
          }
        ],
        "tags": [
          "Login"
        ],
        "summary": "Login",
        "responses": {
          "200": {
            "description": "OK"
          },
          "403": {
            "description": "FORBIDDEN"
          }
        }
      }
    },
    "/register/api": {
      "post": {
        "tags": [
          "Register User"
        ],
        "summary": "User Registration",
        "requestBody": {
          "description": "Post New User To Database",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Users"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "NOT FOUND"
          }
        }
      }
    },
    "/courses/advanced/api": {
      "get": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Get advanced cource"
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "NOT FOUND"
          }
        }
      }
    },
    "/courses/fundamental/api": {
      "get": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Get advanced cource"
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "NOT FOUND"
          }
        }
      }
    },
    "/add_course/api": {
      "post": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Add New Course"
        ],
        "summary": "Add New Course",
        "requestBody": {
          "description": "Add new Course",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Add_Course"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "NOT FOUND"
          }
        }
      }
    },
    "/course/{course_id}/update/": {
      "patch": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Update course by given course_id"
        ],
        "summary": "Update Course",
        "parameters": [
          {
            "in": "path",
            "required": true,
            "name": "course_id",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "requestBody": {
          "description": "Update Course",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Update_Course"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "NOT FOUND"
          }
        }
      }
    },
    "/courses/course/{course_id}": {
      "delete": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Delete course by given course_id"
        ],
        "summary": "Delete Course",
        "parameters": [
          {
            "in": "path",
            "required": true,
            "name": "course_id",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "NOT FOUND"
          }
        }
      }
    },
    "/balance/api/{user}": {
      "get": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Get User Account Balance"
        ],
        "parameters": [
          {
            "in": "path",
            "required": true,
            "name": "user",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "NOT FOUND"
          }
        }
      }
    },
    "/buy_course/api/{id}/{user}": {
      "post": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Buy Course"
        ],
        "summary": "Buy Course by ID",
        "parameters": [
          {
            "in": "path",
            "required": true,
            "name": "id",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "path",
            "required": true,
            "name": "user",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "NOT FOUND"
          }
        }
      }
    },
    "/add_account_balance/{user}/api": {
      "post": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Add Amount"
        ],
        "summary": "Add Amount",
        "parameters": [
          {
            "in": "path",
            "required": true,
            "name": "user",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Add Amount",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Amount"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "NOT FOUND"
          }
        }
      }
    }
  }
}